requirements.txt
streamlit==1.21.0
pandas==1.5.3
opencv-python==4.7.0.68
numpy==1.24.2
Pillow==9.5.0
ultralytics==8.0.18
streamlit-webrtc==0.33.0
torch==2.1.0
